[api minor] Add alias exportTo() to export()
`export` is a reserved word in JavaScript, so it is illegal to use it as
an identifier. Many linting tools will complain about a regular call to
export(module), requiring the following (ugly) fix:

-    ).export(module);
+    )['export'](module);

To encourage good style while maintaining backwards compatibility,
the exportTo(module) alias can be used as a syntactically valid
replacement for export(module).

Closely related issue/fix in vows:

https://github.com/cloudhead/vows/issues/27
https://github.com/cloudhead/vows/commit/398443dab28d4ed3aa313bdf22915df097a9a451